refactor: change header structure
