--- src/components/Header/headerForm.tsx
@@ -1,12 +1,0 @@
-
-
-export const HeaderForm: React.FC = () => {
-
-    return(
-        <>
-            <form data-id="search-form" className="header-controls-search-form form-inline invisible">
-                <input className="form-control" placeholder="Поиск" />
-            </form>
-        </>
-    )
-}
