+++ src/components/pages/MainPage/mainConteiner.tsx
@@ -9,6 +9,7 @@ export const MainConteiner: React.FC = () => {
                 <div className="row">
                     <div className="col">
                         <Banner />
+                        
                     </div>
                 </div>
             </main>
