+++ src/components/pages/MainPage/mainPage.tsx
@@ -1,4 +1,4 @@
-import { Header } from "../../Header/Header"
+import { Header } from "../../Header"
 import { MainConteiner } from "./mainConteiner"
